test: merge amplitude YAML tests into test_yaml
